Fix describe to work without SessionContext
